+++ src/Logic/Line.cpp
@@ -12,14 +12,14 @@ void Line::calculateFactorB(const Point &first, const Point &second) {
     b = first.getY() - first.getX()*onlyCalculateA(first, second);
 }
 
-double Line::onlyCalculateA(const Point &first, const Point &second) const {
+float Line::onlyCalculateA(const Point &first, const Point &second) const {
     if(first.sameXCoordinate(second)){
         throw "X coordinates are the same for those points";
     }
     return (first.getY() - second.getY())/(first.getX() - second.getX());
 }
 
-double Line::calculate(double x) {
+float Line::calculate(float x) {
     return a*x + b;
 }
 
@@ -28,11 +28,11 @@ Line::Line(Point &first, Point &second) {
     calculateFactorB(first, second);
 }
 
-double Line::getA() const {
+float Line::getA() const {
     return a;
 }
 
-double Line::getB() const {
+float Line::getB() const {
     return b;
 }
